continuar

+++ app/(dashboard)/relatorio/actions.ts
@@ -8,20 +8,28 @@ import type {
   ReportRelType,
   ReportUpdateType,
 } from '@/lib/types'
-import { uploadObject } from '@/service/storage'
+import { deleteObject, uploadObject } from '@/service/storage'
 import type { Client } from '@prisma/client'
+import sharp from 'sharp'
 
 export async function upLoadPhotoAnalisys(formData: FormData, idReport: number) {
+  if (!formData || !idReport) {
+    throw new Error('Parâmetros inválidos')
+  }
   const file = formData.get('file') as File
 
-  const binaryFile = await file.arrayBuffer()
-  const fileBuffer = Buffer.from(binaryFile)
-
-  const keyName = `reports/${idReport}/${file.name}`
-
-  const result = uploadObject(keyName, fileBuffer, file)
-
-  return result
+  try {
+    const binaryFile = await file.arrayBuffer()
+    // resize image to 430x280 and convert to jpeg
+    const resizedImage = await sharp(binaryFile).resize(430, 300).toFormat('jpeg').jpeg({ quality: 80 }).toBuffer()
+    const fileBuffer = Buffer.from(resizedImage)
+    const keyName = `reports/${idReport}/${file.name}`
+    const result = await uploadObject(keyName, fileBuffer, file)
+    return result
+  } catch (error) {
+    console.error('Erro ao processar imagem:', error)
+    throw new Error('Erro ao processar imagem')
+  }
 }
 
 export async function savePhotoAnalisys(data: PhotoAnalisysType) {
@@ -65,7 +73,15 @@ export async function getPhotoAnalisysById(id: number) {
   return photoAnalisys
 }
 
-export async function deletePhotoAnalisys(id: number) {
+export async function deletePhotoAnalisys(photoAnalisysData: PhotoAnalisysType) {
+  const id = photoAnalisysData.id
+
+  const result = deleteObject(photoAnalisysData.url)
+
+  if (!result) {
+    throw new Error('Erro ao deletar imagem')
+  }
+
   return await prisma.photoAnalisys.delete({
     where: {
       id,
